+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- `TestCoroutineDispatcherRule` and `TestCoroutineDispatcherExtension` provide `TestCoroutineScheduler` to the tests
+
 ### Changed
+- Replace usages of `TestCoroutineDispatcher` with `UnconfinedTestDispatcher`
 - Upgrade Gradle to `7.4.2`
 - Dependencies:
   - Upgrade Kotlin to `1.6.20`